+++ src/load_stub_files.ts.js
@@ -4,18 +4,31 @@ function load_stub_files(context, output_channel) {
     return async () => {
 
         const adams_stub_dir = context.asAbsolutePath('resources/adamspy');
-        // const adams_stub_files = fs.readdirSync(adams_stub_dir);
         var extra_paths = vscode.workspace.getConfiguration('python').get('analysis.extraPaths', null);
-        // If the stub directory is not already in the extra paths, add it
+        var autocomp_paths = vscode.workspace.getConfiguration('python').get('autoComplete.extraPaths', null);
+        
+        // If the stub directory is not already in the analysis.extraPaths, add it
         if (!extra_paths.includes(adams_stub_dir)) {
             extra_paths.push(adams_stub_dir);
-            output_channel.appendLine(`[${new Date().toLocaleTimeString()}]: Adding "${adams_stub_dir}" to python.autoComplete.extraPaths`);
+            output_channel.appendLine(`[${new Date().toLocaleTimeString()}]: Adding "${adams_stub_dir}" to python.analysis.extraPaths`);
         }
         if (vscode.workspace.workspaceFolders == undefined) {
             vscode.workspace.getConfiguration('python').update('analysis.extraPaths', extra_paths, true);
         }
         else {
             vscode.workspace.getConfiguration('python').update('analysis.extraPaths', extra_paths, null);
+        }
+
+        // If the stub directory is not already in the autoComplete.extra paths, add it
+        if (!autocomp_paths.includes(adams_stub_dir)) {
+            autocomp_paths.push(adams_stub_dir);
+            output_channel.appendLine(`[${new Date().toLocaleTimeString()}]: Adding "${adams_stub_dir}" to python.autoComplete.extraPaths`);
+        }
+        if (vscode.workspace.workspaceFolders == undefined) {
+            vscode.workspace.getConfiguration('python').update('autoComplete.extraPaths', autocomp_paths, true);
+        }
+        else {
+            vscode.workspace.getConfiguration('python').update('autoComplete.extraPaths', autocomp_paths, null);
         }
 
     };